Error.Is pr suggestion
